\name{Maxwind}
\alias{Maxwind}
\docType{data}
\title{Maximum wind speeds}
\description{
Annual maximum wind speeds at 12 sites in the southeast U.S.
}
\usage{data(Maxwind)}
\format{
A list of 12 numeric vectors.
}
\details{
The name of a list element is the site location, including
a reference number used by Simiu et al. (1979).
Each list element is a numeric vector containing the annual
maximum wind speeds for that site.
The period of observation varies from site to site:
for details see Simiu et al. (1979).
}
\source{
Simiu, E., Changery, M. J., and Filliben, J. J. (1979).
Extreme wind speeds at 129 stations in the contiguous United States.
\emph{Building Science Series 118}, National Bureau of Standards,
Washington, D.C.

}
%\references{}
\examples{
data(Maxwind)
}
\keyword{datasets}
